requirements.txt
Flask
google-generativeai
langchain
pandas
gtts
deep-translator
Flask-Cors
Flask-SocketIO
gunicorn
